fixed vcf parser
