+++ source/BAM-file-processing/BAM_file_processing_2_AddOrReplaceReadGroups.sge
@@ -1,0 +1,41 @@
+#!/bin/bash -l
+
+# Second step in processing BAM files: Script runs AddOrReplaceReadGroups (GATK) on BAM files. 
+# AddOrReplaceReadGroups parameters: $1 = Input (BAM), $2 = Output (BAM).
+
+# queue in batch or test
+#$ -cwd
+#$ -S /bin/bash
+#$ -o logs/stdout_AddOrReplaceReadGroups.txt
+#$ -e logs/stderr_AddOrReplaceReadGroups.txt
+#$ -pe mpi 2
+#$ -q test.q
+
+
+# Extract library name
+	FIRST=${1#*lib}
+	LIBNAME=${FIRST%%_*}
+
+# Extract read group platform
+	FIRST=${1#*N?-}
+	RGPU=${FIRST%%_*}
+
+# Extract sample name
+	FIRST=${1#*[0-9]_}
+	SECOND=${FIRST%%lib*}
+	SAMPLENAME=${SECOND%_}
+
+
+
+# Run AddOrReplaceReadGroups (GATK)
+
+gatk AddOrReplaceReadGroups \
+-I $1 \
+-O $2 \
+-RGLB $LIBNAME \
+-RGPL ILLUMINA \
+-SORT_ORDER coordinate \
+-RGPU $RGPU \
+-RGSM $SAMPLENAME
+
+
