+++ source/BAM-file-processing/logs/stdout.txt
@@ -1,0 +1,12 @@
+Tool returned:
+SUCCESS
+Tool returned:
+SUCCESS
+Tool returned:
+SUCCESS
+Tool returned:
+SUCCESS
+Tool returned:
+SUCCESS
+Tool returned:
+SUCCESS
