+++ source/BAM-file-processing/logs/stdout_AddOrReplaceReadGroups.txt
@@ -1,0 +1,12 @@
+Tool returned:
+0
+Tool returned:
+0
+Tool returned:
+0
+Tool returned:
+0
+Tool returned:
+0
+Tool returned:
+0
