+++ source/BAM-file-processing/manual_scripts/BQSR-Picard_GATK.sge
@@ -1,0 +1,55 @@
+#!/bin/bash -l
+
+# Script runs base quality score recalibration (BQSR) through GATK. 
+# Script defined for one bam file. Run on a list of bam files through a loop.
+# Example: 'cat bam_files.txt | while read R; do qsub script $R; done'
+# INPUT: BAM.
+# OUTPUT: BAM
+
+# queue in batch
+#$ -cwd
+#$ -S /bin/bash
+#$ -pe mpi 5
+#$ -q batch.q
+#$ -o logs/stdout.txt
+#$ -e logs/stderr.txt
+
+
+# Rename input variable
+R=$1
+
+# Set file variables 
+
+	# Set variable for path to read files
+	WORK_DIR=/home/xlorda/anna_tmp/mapped_bam_files
+
+	# Set variable to reference genome.fasta
+	HG38=/home/xlorda/anna_tmp/reference_and_misc_files/GRCh38.primary_assembly.genome.fa	
+
+	# Set variable to known sites SNPS.VCF
+	KNOWN_SNPS=/home/xlorda/anna_tmp/reference_and_misc_files/Homo_sapiens_assembly38.dbsnp138.vcf
+
+	# Set variable to known sites INDELS.VCF
+	KNOWN_INDELS=/home/xlorda/anna_tmp/reference_and_misc_files/Mills_and_1000G_gold_standard.indels.hg38.vcf.gz
+
+	# Set variable for table file, output from BQSR, input for apply BQSR.
+	TABLE_FILE=${WORK_DIR}/BQSR_table_${R}
+
+
+# Generate recalibration table for each input bam
+gatk BaseRecalibrator \
+-R $HG38 \
+-I $WORK_DIR/$R \
+--known-sites $KNOWN_SNPS \
+--known-sites $KNOWN_INDELS \
+-O $TABLE_FILE
+
+
+
+# Apply model generated by BQSR to adjust base quality scores, outputs new BAM file
+gatk ApplyBQSR \
+-R $HG38 \
+-I $WORK_DIR/$R \
+--bqsr-recal-file $TABLE_FILE \
+-O ${WORK_DIR}/BQSR_${R}
+
