+++ source/BAM-file-processing/manual_scripts/picard_Mark_Duplicates.sge
@@ -1,0 +1,26 @@
+#!/bin/bash -l
+
+# Script defined for one bam file. Run on a list of bam files through a loop.
+# Example: 'cat bam_files.txt | while read R; do qsub script $R; done'
+
+# queue in batch
+#$ -cwd
+#$ -S /bin/bash
+#$ -pe mpi 10
+#$ -q batch.q
+
+R=$1
+
+#Set variable for path to read files
+FILE_PATH=/seqstore/remote/share/anna_tmp/mapped_bam_files
+
+# Create prefix that will be added to output files
+OUT_PREFIX=RMDUP
+
+# Run Picard tools MarkDuplicates through GATK
+
+gatk MarkDuplicates \
+-I $FILE_PATH/${R} \
+-O $FILE_PATH/${OUT_PREFIX}_${R} \
+-M $FILE_PATH/${R}_metrics.txt
+
