--- source/BAM_file_processing/BAM_file_processing_3_BaseRecalibrator.sge
@@ -1,22 +1,0 @@
-#!/bin/bash -l
-
-# Third step in processing BAM files: Script runs BaseRecalibrator (GATK) on BAM files, generating a base recalibration model (table) for each BAM file input. 
-# BaseRecalibrator parameters: $1 = Input (BAM), $2 = Output (base recalibration model, table file), $3 = Reference genome (HG38), $4 = Known SNPs file, $5 = Known Indels file.
-
-# queue in batch or test
-#$ -cwd
-#$ -S /bin/bash
-#$ -pe mpi 5
-#$ -q test.q
-#$ -o logs/stdout.txt
-#$ -e logs/stderr.txt
-
-# Run BaseRecalibrator (GATK) 
-
-gatk BaseRecalibrator \
--I $1 \
--O $2 \
--R $3 \
---known-sites $4 \
---known-sites $5
-
