--- source/BAM_file_processing/logs/stdout.txt
@@ -1,12 +1,0 @@
-Tool returned:
-SUCCESS
-Tool returned:
-SUCCESS
-Tool returned:
-SUCCESS
-Tool returned:
-SUCCESS
-Tool returned:
-SUCCESS
-Tool returned:
-SUCCESS
