--- source/BAM_file_processing/logs/stdout_AddOrReplaceReadGroups.txt
@@ -1,12 +1,0 @@
-Tool returned:
-0
-Tool returned:
-0
-Tool returned:
-0
-Tool returned:
-0
-Tool returned:
-0
-Tool returned:
-0
