--- source/BAM_file_processing/manual_scripts/Mark_Duplicates-Picard_GATK.sge
@@ -1,29 +1,0 @@
-#!/bin/bash -l
-
-# Script runs Picard tools Mark Duplicates function (through gatk) on BAM files. 
-# Script defined for one bam file. Run on a list of bam files through a loop.
-# Example: 'cat bam_files.txt | while read R; do qsub script $R; done'
-# INPUT: BAM. 
-# OUTPUT: BAM. 
-
-# Queue in batch
-#$ -cwd
-#$ -o logs/stdout.txt
-#$ -e logs/stderr.txt
-#$ -S /bin/bash
-#$ -pe mpi 10
-#$ -q batch.q
-
-# Set variable for input
-R=$1
-
-# Set variable for work dir
-WORK_DIR=/home/xlorda/anna_tmp/mapped_bam_files
-
-# Run Picard tools MarkDuplicates through GATK
-
-gatk MarkDuplicates \
--I ${WORK_DIR}/${R} \
--O ${WORK_DIR}/RMDUP_${R} \
--M $WORK_DIR/${R}_metrics.txt
-
