+++ source/config.txt
@@ -1,0 +1,53 @@
+# Config file for: BAM-file processing, Panel of Normals (PoN) generation, and Somatic variant calling.
+
+
+# Common variables
+
+        WORK_DIR=/home/xlorda/anna_tmp/test_rerun:
+
+
+	# Reference files: 
+
+        REF_DIR=/home/xlorda/anna_tmp/reference_and_misc_files
+
+        HG38=${REF_DIR}/GRCh38.primary_assembly.genome.fa
+
+        KNOWN_SNPS=${REF_DIR}/Homo_sapiens_assembly38.dbsnp138.vcf
+
+        KNOWN_INDELS=${REF_DIR}/Mills_and_1000G_gold_standard.indels.hg38.vcf.gz
+
+	INTERVALS_LIST=${REF_DIR}/wgs_calling_regions.hg38.list
+
+	GERMLINE_RESOURCE=${REF_DIR}/af-only-gnomad.hg38.vcf.gz
+
+
+
+
+# BAM file processing variables:
+
+        BAM_SCRIPT_PATH=/home/xlorda/Biomarkers-immuno-lungcancer/source/BAM_file_processing
+
+        SamtoolsIndex=${BAM_SCRIPT_PATH}/BAM_file_processing_0_Samtools_index.sge
+
+        MarkDuplicates=${BAM_SCRIPT_PATH}/BAM_file_processing_1_MarkDuplicates.sge
+
+        AddOrReplaceReadGroups=${BAM_SCRIPT_PATH}/BAM_file_processing_2_AddOrReplaceReadGroups.sge
+
+        BaseRecalibrator=${BAM_SCRIPT_PATH}/BAM_file_processing_3_BaseRecalibrator.sge
+
+        ApplyBQSR=${BAM_SCRIPT_PATH}/BAM_file_processing_4_ApplyBQSR.sge
+
+
+
+# PoN generation variables:
+
+        PON1=/home/xlorda/Biomarkers-immuno-lungcancer/source/PoN/PoN1.sge
+
+        PON2=/home/xlorda/Biomarkers-immuno-lungcancer/source/PoN/PoN2.sge
+
+        PON3=/home/xlorda/Biomarkers-immuno-lungcancer/source/PoN/PoN3.sge
+
+	PON_DB=${WORK_DIR}/PON_DB
+
+
+
